Шаг полностью сделанный + немного от 5го 4.03

+++ app/Http/Controllers/TaskStatusController.php
@@ -79,7 +79,12 @@ class TaskStatusController extends Controller
     public function destroy(string $id)
     {
         //тут будет проверка на наличие удаляемого статуса у задачи, если установлен, то не удаляем
+        $taskStatus = TaskStatus::findOrFail($id);
 
+        if ($taskStatus->tasks()->exists()) {
+            flash()->error('Не удалось удалить статус');
+            return redirect(route('task_statuses.index'));
+        }
         TaskStatus::destroy($id);
         flash()->info('Статус удален!');
         return redirect(route('task_statuses.index'));
